feat(execute): Add Engine RPC option to `execute remote` (#2070)
* feat(rpc): Allow EngineRPC to have a custom jwt secret

* fix(rpc): Allow specifying a namespace for RPC subclasses

* refactor(rpc): Small changes

* feat(execute): Allow to specify an Engine RPC endpoint to drive the chain

refactor(execute): rename class

fix(plugins/execute): rebase issues

* docs: Update `docs/running_tests/execute/remote.md`

* docs: changelog
