+++ ethereum_test_rpc/rpc.py
@@ -57,23 +57,22 @@ class BaseRPC:
     def __init__(
         self,
         url: str,
-        extra_headers: Dict | None = None,
+        *,
         response_validation_context: Any | None = None,
     ):
         """Initialize BaseRPC class with the given url."""
-        if extra_headers is None:
-            extra_headers = {}
         self.url = url
         self.request_id_counter = count(1)
-        self.extra_headers = extra_headers
         self.response_validation_context = response_validation_context
 
-    def __init_subclass__(cls) -> None:
+    def __init_subclass__(cls, namespace: str | None = None) -> None:
         """Set namespace of the RPC class to the lowercase of the class name."""
-        namespace = cls.__name__
-        if namespace.endswith("RPC"):
-            namespace = namespace.removesuffix("RPC")
-        cls.namespace = namespace.lower()
+        if namespace is None:
+            namespace = cls.__name__
+            if namespace.endswith("RPC"):
+                namespace = namespace.removesuffix("RPC")
+            namespace = namespace.lower()
+        cls.namespace = namespace
 
     def post_request(
         self,
@@ -99,7 +98,7 @@ class BaseRPC:
         base_header = {
             "Content-Type": "application/json",
         }
-        headers = base_header | self.extra_headers | extra_headers
+        headers = base_header | extra_headers
 
         response = requests.post(self.url, json=payload, headers=headers)
         response.raise_for_status()
@@ -125,18 +124,12 @@ class EthRPC(BaseRPC):
 
     def __init__(
         self,
-        url: str,
-        extra_headers: Dict | None = None,
-        *,
+        *args,
         transaction_wait_timeout: int = 60,
-        response_validation_context: Any | None = None,
+        **kwargs,
     ):
         """Initialize EthRPC class with the given url and transaction wait timeout."""
-        if extra_headers is None:
-            extra_headers = {}
-        super().__init__(
-            url, extra_headers, response_validation_context=response_validation_context
-        )
+        super().__init__(*args, **kwargs)
         self.transaction_wait_timeout = transaction_wait_timeout
 
     def config(self):
@@ -331,6 +324,18 @@ class EngineRPC(BaseRPC):
     simulators.
     """
 
+    jwt_secret: bytes
+
+    def __init__(
+        self,
+        *args,
+        jwt_secret: bytes = b"secretsecretsecretsecretsecretse",  # Default secret used in hive
+        **kwargs,
+    ):
+        """Initialize Engine RPC class with the given JWT secret."""
+        super().__init__(*args, **kwargs)
+        self.jwt_secret = jwt_secret
+
     def post_request(
         self,
         method: str,
@@ -343,7 +348,7 @@ class EngineRPC(BaseRPC):
             extra_headers = {}
         jwt_token = encode(
             {"iat": int(time.time())},
-            b"secretsecretsecretsecretsecretse",  # the secret used within clients in hive
+            self.jwt_secret,
             algorithm="HS256",
         )
         extra_headers = {
